requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
Flask-OAuth==0.12
Pillow==9.4.0
opencv-python==4.7.0.72
pandas==1.5.3
numpy==1.24.2
gunicorn==20.1.0
Werkzeug==2.2.3
email-validator==1.3.1
python-dotenv==1.0.0
itsdangerous==2.1.2
Authlib==1.2.0
requests==2.28.2
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
tensorflow==2.9.1
keras==2.9.0
deepface==0.0.79
